fix[#7] : 루트카테고리가 삭제되면 해당 카테고리를 가지고 있는 모든 shop 삭제
- 효율적인 쿼리를 위해 categoryService에서 deleteAllInBatch 를 쓰고 싶었지만 참조무결성 오류가 자꾸 발생하여 사용하지 못하였다.... (찾는중)

+++ backend/src/main/java/com/handong/rebon/category/domain/Category.java
@@ -75,17 +75,15 @@ public class Category extends BaseEntity {
         if (isDeleted()) {
             throw new CategoryAlreadyDeletedException();
         }
-        boolean isParentCategory = Objects.isNull(this.parent);
-        deleteShopCategory(isParentCategory);
-
-        this.children.delete();
-        deleteContent();
+        if (isParentCategory()) {
+            this.children.delete();
+        }
+        this.shopCategories.forEach(BaseEntity::deleteContent);
+        this.deleteContent();
     }
 
-    public void deleteShopCategory(boolean isRootCategory) {
-        this.shopCategories.forEach(shopCategory -> {
-            shopCategory.delete(isRootCategory);
-        });
+    public boolean isParentCategory() {
+        return Objects.isNull(this.parent);
     }
 
     public void addShopCategory(ShopCategory shopCategory) {
